Move away telemetry endpoint argument
