+++ client/telemetry/src/lib.rs
@@ -109,8 +109,7 @@ fn url_or_multiaddr_deser<'de, D>(deserializer: D) -> Result<Vec<(Multiaddr, u8)
 {
 	Vec::<(String, u8)>::deserialize(deserializer)?
 		.iter()
-		.map(|e| Ok((url_to_multiaddr(&e.0)
-		.map_err(serde::de::Error::custom)?, e.1)))
+		.map(|e| Ok((url_to_multiaddr(&e.0).map_err(serde::de::Error::custom)?, e.1)))
 		.collect()
 }
